--- a/base/task.jl
+++ b/base/task.jl
@@ -607,7 +607,7 @@ function wait()
     W = Workqueues[Threads.threadid()]
     reftask = poptaskref(W)
     result = try_yieldto(ensure_rescheduled, reftask)
-    process_events()
+    Sys.isjsvm() || process_events()
     # return when we come out of the queue
     return result
 end