@Keno 2019-07-09
Current status: Things basically work with one task at a time, but when I try to interleave tasks it gets confused. E.g.:
```
julia> c = 0
0
julia> make_task() = (global c; c += 1; let i = c; @task begin; println("Before $i"); ccall(:emscripten_sleep, Cvoid, (UInt32,), 1000); println(current_task()); println("Middle $i"); ccall(:emscripten_sleep, Cvoid, (UInt32,), 1000); println("Hello $i"); end; end
julia> t1 = make_task(); t2 = make_task(); t3 = make_task()
julia> println(t1); println(t2); println(t3);
Task (runnable) @0x03c3e8f0
Task (runnable) @0x03c3ee30
Task (runnable) @0x03c3f370
julia> schedule(t1); schedule(t2); schedule(t3);
Before 1
Before 2
Before 3
Task (runnable) @0x03c3e8f0
Middle 3
Hello 3
```
so it's getting the right task, but printing the wrong `i`. I suspect `i` here is on the c stack in an interpreter frame, so we're probably not restoring those correctly.