@kripken 2019-07-09
One thing the Emscripten C implementation is missing (for coroutines) is saving and restoring the C stack, could be part of the same issue here (Bysyncify itself isn't aware of the C stack, just the wasm native call stack). `stackSave/stackRestore` can be used in JS to get and set the C stack position.