@Keno 2019-07-09
Yes, I think I'm restoring that here (after saving it on the corresponding deschedule routine):
https://github.com/JuliaLang/julia/pull/32532/files#diff-15aaf3bb4e0956d73f314b03ddf85c8cR92

could easily be doing it wrong though. I'll look into it some more.