@Keno 2019-07-10
I did a stupid in the task management code that caused the coroutine to release the stack too early. The example above works now:

<img width="246" alt="Screen Shot 2019-07-09 at 11 15 47 PM" src="https://user-images.githubusercontent.com/1291671/60937793-9da75600-a29f-11e9-9cc7-6f07f77e9bb9.png">