@Keno 2019-07-10
This seems to be working well now.