@Keno 2019-07-11
For people following along the status here is that this works well, though the bysyncify pass still seems to have a significant amount of overhead for our use case (as well as causing Chrome to complain about it's stack being exceeded). I'm fairly confident that that can be addressed by being a bit more careful which functions are transformed.