@kripken 2019-07-11
How much overhead are you seeing? Aside from the issue of picking which functions to transform (which I saw you opened an issue on, good idea!), I'd be surprised if this increases code size more than say 50% or so, and especially I'm surprised it overflows the native call stack in browsers. Maybe you can send me a testcase to look at?