@Keno 2019-07-12
> Maybe you can send me a testcase to look at?

Alright, I've pushed a kf/bysyncifybench branch to https://github.com/Keno/julia-wasm. The hello.js and hello.wasm files in that directory are the build outputs with bysyncify enabled. The -no-bysyncify variants of those files are as the name implies. Local benchmarks:

Bysyncify expands the .wasm file by 3.4x. Time benchmarks:

```
GC.enable(false) # Turn off GC for benchmarking, also we may still have GC bugs on wasm
show(stdout, MIME"text/html", methods(gcd)) # Run something that wasn't compiled yet, triggering the compiler to be run (very, very slow at the moment on wasm)
```

Without Bysyncify:
- Firefox Nightly (IonJit): 22s
- Firefox Nightly (Baseline): 29s
- Chrome 75: 26s

With Bysyncify:
- Firefox Nightly (IonJit): Becomes memory bomb (https://bugzilla.mozilla.org/show_bug.cgi?id=1561127)
- Firefox Nightly (Baseline): 444s
- Chrome: Runs out of stack

The julia runtime itself was compiled with O3 (which is what the time benchmarks are measuring), but I re-used dependencies from an O0 build, so some of the size expansion may be easily explained by that.