@kripken 2019-07-12
Thanks @Keno! 

Optimizations do seem to explain most of the code size issues here:

 * Optimizing `hello-no-bysyncify.wasm` with wasm-opt makes it 15% smaller, which suggests it wasn't optimized by wasm-opt before (emscripten would run it automatically for you, but maybe some files came from somewhere else?).
 * Bysyncifying it without optimizations leads to a similarly very large result as in the `hello.wasm` file there.
 * Bysyncifying it **with** optimizations, using `wasm-opt -O --bysyncify hello-no-bysyncify.wasm -o c.wasm` (which is what emscripten would do if optimizing) makes it 3.1x smaller, and just 25% larger than the (optimized) `hello-no-bysyncify.wasm`, which is what I'd expect.

Based on my own testing, the size and speed overheads tend to be correlated, so I'm hopeful that building with optimizations will give a much smaller slowdown - hopefully just around 25% or so.

(I wanted to try to measure speed myself, but wasn't sure how? When I open a webserver to `website/` I get a login and a button to "try Julia now" but the button seems to just error.)