@Keno 2019-07-12
> Optimizing hello-no-bysyncify.wasm with wasm-opt makes it 15% smaller, which suggests it wasn't optimized by wasm-opt before (emscripten would run it automatically for you, but maybe some files came from somewhere else?).

Ugh. That .wasm file did come right out of emscripten, but I think I forgot to put the `-O` on the final emcc link line that produces the .js, so it probably didn't run wasm-opt because of that. Let me redo the measurement with `-O` on that link line.

> (I wanted to try to measure speed myself, but wasn't sure how? When I open a webserver to website/ I get a login and a button to "try Julia now" but the button seems to just error.)

Sorry, the frontend is a hacked up version of a frontend we had many years ago and is super janky (I just threw it together in 20 minutes). The place to go to is `repl.html`, which'll load the hello.js/hello.wasm file.