@Keno 2019-07-12
Ok, so with optimizations, I do see the small file size for the .wasm file, but unfortunately the timings are still in the ~350s range. In more encouraging news though it doesn't blow up the browsers anymore, so that's progress.