@Keno 2019-07-12
Interestingly Chrome seems to be quite a bit better (now that the stack doesn't blow up anymore), clocking in at 81s on the optimized .wasm.