@kripken 2019-07-12
Thanks @Keno, now I see the repl and I can enter commands, but what do I enter to run the benchmark? (I tried to enter the two lines I saw before, but the first already errors, `GC.enable(false)` leads throws `js_run_pending_tasks is not a function`).

(81s vs 26s (over 3x slower) is surprisingly bad, so I'd like to do some profiling to investigate this.)