@Keno 2019-07-12
> js_run_pending_tasks is not a function

Sorry about that, forgot to push a change. Try with https://github.com/Keno/julia-wasm/commit/52944b469376fa2041832b37b60a485e7eec072f (on the same branch).